--- modules/cucumber/modules/codegen/src/main/resources/META-INF/services/io.swagger.codegen.CodegenConfig
@@ -1,1 +1,0 @@
-com.smartbear.readyapi.testserver.codegen.TestServerCucumberFeatureGenerator
