--- modules/cucumber/modules/codegen/src/main/resources/TestServerCucumberCodeGen/feature.mustache
@@ -1,11 +1,0 @@
-Feature: {{appDescription}} - Version {{version}}
-
-Background:
-  Given the Swagger definition at {{basePath}}swagger.json
-{{#operations}}{{#operation}}{{#responses}}
-  Scenario: {{description}}
-    When a request to {{operationId}} is made
-    {{#allParams}}And {{baseName}} is {{defaultValue}}
-    {{/allParams}}Then a {{code}} response is returned
-{{/responses}}
-{{/operation}}{{/operations}}
