+++ modules/cucumber/modules/runner/src/docker/resources/assembly.xml
@@ -6,7 +6,7 @@
             <useProjectArtifact>true</useProjectArtifact>
             <fileMode>444</fileMode>
             <includes>
-                <include>com.smartbear.readyapi.testserver.cucumber:testserver-cucumber-runner</include>
+                <include>com.smartbear.readyapi:readyapi4j-cucumber-runner</include>
             </includes>
             <outputDirectory>.</outputDirectory>
         </dependencySet>
