+++ modules/cucumber/modules/runner/src/docker/resources/entry-point.sh
@@ -4,7 +4,7 @@
 #
 
 # find jar to run
-JAR=$(find -name "*testserver-cucumber*.jar" -type f)
+JAR=$(find -name "*readyapi4j-cucumber*.jar" -type f)
 
 # extract additional JVM options from ENV vars
 OPTS=()
